refactor(orm): Achieve a more generic repository system
Repositories are now provided using a CRUD interface. The only mechanics required is bridging the gap between Entity and ORMModels.
To do so, a Presenter has been introduced, which is internally used by repositories to make the convertion to and from Entities/ORMModels.
Left to the system is the convertion between Entity / ORMModel and vice-versa, which is now easily achieved through the Converter interface.
Each converter are required to be registered on the presenter used by repositories.